+++ cpp/tests/io/parquet_common.hpp
@@ -211,6 +211,13 @@ std::unique_ptr<cudf::column> make_parquet_string_list_col(std::mt19937& engine,
 template <typename T>
 std::pair<cudf::table, std::string> create_parquet_typed_with_stats(std::string const& filename);
 
+int32_t compare_binary(std::vector<uint8_t> const& v1,
+                       std::vector<uint8_t> const& v2,
+                       cudf::io::parquet::detail::Type ptype,
+                       thrust::optional<cudf::io::parquet::detail::ConvertedType> const& ctype);
+
+void expect_compression_stats_empty(std::shared_ptr<cudf::io::writer_compression_statistics> stats);
+
 // =============================================================================
 // ---- test data for stats sort order tests
 // need at least 3 pages, and min page count is 5000, so need at least 15000 values.
